[#6] refactor: redis store 분리
- cache 기술과 관련된 로직을 store가 관리
